# openapi.yaml
basePath: /
consumes:
- application/json
- multipart/form-data
...
host: 127.0.0.1:5000
info:
  contact:
    name: Plant Disease Detection Team
    url: https://github.com/yourusername/PlantDisease
  description: REST API for plant disease detection with MongoDB storage. Supports
    multi-language plant/disease information and user-contributed names.
  title: Plant Disease Detection API
  version: 0.1.0
paths:
...
  /api/plants:
    get:
      responses:
        '200':
          description: List of all plants
          schema:
            items:
              properties:
                _id:
                  type: string
                canonical_name:
                  type: string
                care_instructions:
                  type: object
                description:
                  type: object
                names:
                  type: object
                scientific_name:
                  type: string
                user_contributed_names:
                  type: array
              type: object
            type: array
      summary: Get all plants
      tags:
      - Plants
    post:
      consumes:
      - application/json
      parameters:
      - in: body
        name: body
        required: true
        schema:
          properties:
            canonical_name:
              example: Tomato
              type: string
            care_instructions:
              description: Multi-language care instructions
              example:
                ar: Water_regularly_in_Arabic
                en: Water regularly
              type: object
            description:
              description: Multi-language descriptions
              example:
                en: Common garden vegetable
              type: object
            names:
              description: Multi-language names
              example:
                ar:
                - Tomato_in_Arabic
                en:
                - Tomato
                fr:
                - Tomate
              type: object
            scientific_name:
              example: Solanum lycopersicum
              type: string
          required:
          - canonical_name
          type: object
      responses:
        '201':
          description: Plant created successfully
          schema:
            properties:
              id:
                type: string
            type: object
        '400':
          description: Invalid request data
      summary: Create new plant
      tags:
      - Plants
  /api/plants/{plant_id}:
    delete:
      parameters:
      - description: Plant ID
        in: path
        name: plant_id
        required: true
        type: string
      responses:
        '200':
          description: Plant deleted successfully
          schema:
            properties:
              message:
                type: string
            type: object
        '400':
          description: Invalid plant ID
        '404':
          description: Plant not found
      summary: Delete plant
      tags:
      - Plants
    get:
      parameters:
      - description: Plant ID
        in: path
        name: plant_id
        required: true
        type: string
      responses:
        '200':
          description: Plant details
          schema:
            properties:
              _id:
                type: string
              canonical_name:
                type: string
              care_instructions:
                description: Multi-language care instructions
                type: object
              description:
                description: Multi-language descriptions
                type: object
              names:
                description: Multi-language names
                example:
                  ar:
                  - Tomato_in_Arabic
                  en:
                  - Tomato
                type: object
              scientific_name:
                type: string
              user_contributed_names:
                items:
                  properties:
                    added_by:
                      type: string
                    created_at:
                      format: date-time
                      type: string
                    language:
                      type: string
                    name:
                      type: string
                    votes:
                      type: integer
                  type: object
                type: array
            type: object
        '400':
          description: Invalid plant ID
        '404':
          description: Plant not found
      summary: Get plant by ID
      tags:
      - Plants
    put:
      consumes:
      - application/json
      parameters:
      - description: Plant ID
        in: path
        name: plant_id
        required: true
        type: string
      - in: body
        name: body
        required: true
        schema:
          properties:
            canonical_name:
              type: string
            care_instructions:
              type: object
            description:
              type: object
            names:
              type: object
            scientific_name:
              type: string
          type: object
      responses:
        '200':
          description: Plant updated successfully
          schema:
            properties:
              message:
                type: string
            type: object
        '400':
          description: Invalid request data
        '404':
          description: Plant not found
      summary: Update plant
      tags:
      - Plants
  /api/plants/{plant_id}/names:
    post:
      consumes:
      - application/json
      parameters:
      - description: Plant ID
        in: path
        name: plant_id
        required: true
        type: string
      - in: body
        name: body
        required: true
        schema:
          properties:
            added_by:
              description: User ID who added the name
              example: user_123
              type: string
            language:
              description: Language code (en, ar, fr, etc.)
              example: en
              type: string
            name:
              description: User-contributed name for the plant
              example: Love Apple
              type: string
          required:
          - name
          type: object
      responses:
        '200':
          description: Name added successfully
          schema:
            properties:
              message:
                type: string
            type: object
        '400':
          description: Invalid request data
        '404':
          description: Plant not found
      summary: Add user-contributed plant name
      tags:
      - Plants
  /api/plants/{plant_id}/names/{name_index}/vote:
    post:
      consumes:
      - application/json
      parameters:
      - description: Plant ID
        in: path
        name: plant_id
        required: true
        type: string
      - description: Index of the contributed name in the array
        in: path
        name: name_index
        required: true
        type: integer
      - in: body
        name: body
        required: true
        schema:
          properties:
            vote:
              description: Vote value (1 for upvote, -1 for downvote)
              enum:
              - 1
              - -1
              example: 1
              type: integer
          required:
          - vote
          type: object
      responses:
        '200':
          description: Vote recorded successfully
          schema:
            properties:
              message:
                type: string
            type: object
        '400':
          description: Invalid vote value or plant ID
        '404':
          description: Plant not found
      summary: Vote on user-contributed plant name
      tags:
      - Plants
...
produces:
- application/json
schemes:
- http
swagger: '2.0'